move the map method to the EventsSection function and move types to events-section.tsx
